requirements.txt
cryptography
django==3.2
djangorestframework==3.12.2
django-allauth
dj-rest-auth
psycopg2==2.8.6
psycopg2-binary
Pillow==8.1.0
flake8
djangorestframework-simplejwt==4.6.0
django-cors-headers
django-environ
dj-rest-auth
django-allauth
django-filter
cloudinary
markdown
ulid-py
# for development
# django-debug-toolbar
# drf-yasg==1.20.0
# django-extensions
# django-imagekit
# for prodction
dj_database_url
django_heroku
gunicorn
whitenoise
django-imagekit
django-storages
django-cleanup
boto3
black
autopep8
pipdeptree
#flake8-3.9.1 pycodestyle-2.7.0 pyflakes-2.3.1